requirements.txt
fastapi
uvicorn
influxdb-client
python-dotenv
pyyaml
scapy